+++ packages/common-ethereum/src/index.ts
@@ -3,3 +3,13 @@
 
 export * from './codegen';
 export * from './project';
+
+import {INetworkCommonModule} from '@subql/types-core';
+import {SubqlDatasource, SubqlRuntimeDatasource, SubqlCustomDatasource} from '@subql/types-ethereum';
+import * as p from './project';
+
+// This provides a compiled time check to ensure that the correct exports are provided
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+const _ = {
+  ...p,
+} satisfies INetworkCommonModule<SubqlDatasource, SubqlRuntimeDatasource, SubqlCustomDatasource>;