+++ packages/common-ethereum/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add alias `parseProjectManifest`, also follow type of `INetworkCommonModule` (#318)
+
 ## [3.10.0] - 2024-06-18
 ### Changed
 - Update `@subql/common` dependency